Update sample_data_analysis.py
Best update so far. All that needs to be fixed with the API is pagination and I have asked Naveen to do it. Will fix later but so far we have an api that we can call and get a 10 article long database from!

TO DO We need to check on the data types in the database and ensure that they are actually compatible.